requirements.txt
psycopg2-binary==2.8.3
PyAudio==0.2.11
SpeechRecognition==3.8.1
